[fix] Script additions. Fixed options.yml path to absolute.

+++ compile.sh
@@ -1,6 +1,13 @@
-rm -rf build
+if [ $1 -e "-c" ]
+then
+	rm -rf build
+fi
 mkdir -p build
 cd build
-cmake3 .. -DCMAKE_BUILD_TYPE="Release" .
+if [ $1 -e "-c" ]
+then
+	cmake3 .. -DCMAKE_BUILD_TYPE="Release" .
+fi
 make
-
+rm -rf /opt/manosanag/teraheap-opensrc/allocator/ummap-io/libparallax.so
+cp -r lib/libparallax.so /opt/manosanag/teraheap-opensrc/allocator/ummap-io/
